more cleanup goodness
