+++ core/inferno/iodrv_funcs.c
@@ -26,7 +26,6 @@
 #include <pspumd.h>
 #include <psputilsforkernel.h>
 #include <pspthreadman_kernel.h>
-#include "systemctrl.h"
 #include "systemctrl_se.h"
 #include "systemctrl_private.h"
 #include "inferno.h"
