+++ core/inferno/main.c
@@ -26,7 +26,6 @@
 #include <pspumd.h>
 #include <psprtc.h>
 #include <pspinit.h>
-#include "systemctrl.h"
 #include "systemctrl_se.h"
 #include "systemctrl_private.h"
 #include "inferno.h"
