+++ core/vshctrl/vshpatch.c
@@ -25,9 +25,8 @@
 #include <stdio.h>
 #include <string.h>
 #include <pspumd.h>
-#include "systemctrl.h"
+#include <systemctrl.h>
 #include "xmbiso.h"
-#include "systemctrl.h"
 #include "systemctrl_se.h"
 #include "systemctrl_private.h"
 #include "main.h"
