+++ extras/menus/advancedvsh/main.c
@@ -40,7 +40,6 @@
 
 
 #include "common.h"
-#include "systemctrl.h"
 #include "systemctrl_se.h"
 #include "kubridge.h"
 #include "vpl.h"
