+++ extras/menus/arkMenu/include/entry.h
@@ -6,7 +6,6 @@
 #include <cstdlib>
 #include <vector>
 #include <malloc.h>
-#include "systemctrl.h"
 #include "systemctrl_se.h"
 #include "common.h"
 #include "controller.h"
