+++ extras/modules/xmbctrl/plugins.h
@@ -4,7 +4,6 @@
 #include <pspsdk.h>
 #include <pspkernel.h>
 #include <psputility_sysparam.h>
-#include <systemctrl.h>
 #include <kubridge.h>
 #include <stddef.h>
