+++ extras/modules/xmbctrl/settings.c
@@ -1,7 +1,6 @@
 #include <pspsdk.h>
 #include <pspkernel.h>
 #include <psputility_sysparam.h>
-#include <systemctrl.h>
 #include <kubridge.h>
 #include <stddef.h>
